\name{getStrataLegend-methods}
\docType{methods}
\alias{getStrataLegend-methods}
\alias{getStrataLegend,data.frame,BasicVector-method}

\title{Methods for Function getStrataLegend}
\description{
Get a \code{data.frame} describing the strata of a data set.
}
\section{Methods}{
\describe{

\item{x = "data.frame", design = "BasicVector"}{get a \code{data.frame} 
  describing the strata, according to the variables specified by \code{design}.}
}}
\author{Andreas Alfons, \email{alfons@statistik.tuwien.ac.at}}
\seealso{
  \code{\link{getStrataLegend}}
}
\keyword{methods}
\keyword{manip}
