\name{getStrataSplit-methods}
\docType{methods}
\alias{getStrataSplit-methods}
\alias{getStrataSplit,data.frame,BasicVector-method}

\title{Methods for Function getStrataSplit}
\description{
Get a list in which each element contains the indices of the observations of a 
data set belonging to the corresponding stratum.
}
\section{Methods}{
\describe{

\item{x = "data.frame", design = "BasicVector"}{get a list in which each 
  element contains the indices of the observations belonging to the 
  corresponding stratum, according to the variables specified by \code{design}.}
}}
\author{Andreas Alfons, \email{alfons@statistik.tuwien.ac.at}}
\seealso{
  \code{\link{getStrataSplit}}
}
\keyword{methods}
\keyword{manip}
