\name{getStrataTable-methods}
\docType{methods}
\alias{getStrataTable-methods}
\alias{getStrataTable,data.frame,BasicVector-method}

\title{Methods for Function getStrataTable}
\description{
Get a \code{data.frame} describing the strata and containing the stratum sizes 
of a data set.}
\section{Methods}{
\describe{

\item{x = "data.frame", design = "BasicVector"}{get a \code{data.frame} 
  describing the strata and containing the stratum sizes, according to the 
  variables specified by \code{design}.}
}}
\author{Andreas Alfons, \email{alfons@statistik.tuwien.ac.at}}
\seealso{
  \code{\link{getStrataTable}}
}
\keyword{methods}
\keyword{manip}
