\name{NumericMatrix-class}
\Rdversion{1.1}
\docType{class}
\alias{NumericMatrix-class}

\title{Class "NumericMatrix"}
\description{Virtual class used internally for convenience.}
\section{Objects from the Class}{
  A virtual Class: No objects may be created from it.
}
\section{Methods}{
No methods defined with class "NumericMatrix" in the signature.
}
\author{Andreas Alfons, \email{alfons@statistik.tuwien.ac.at}}
\examples{
showClass("NumericMatrix")
}
\keyword{classes}
