\name{OptCall-class}
\Rdversion{1.1}
\docType{class}
\alias{OptCall-class}

\title{Class "OptCall"}
\description{Virtual class used internally for convenience.}
\section{Objects from the Class}{
  A virtual Class: No objects may be created from it.
}
\section{Methods}{
No methods defined with class "OptCall" in the signature.
}
\author{Andreas Alfons, \email{alfons@statistik.tuwien.ac.at}}
\examples{
showClass("OptCall")
}
\keyword{classes}
