\name{OptCharacter-class}
\Rdversion{1.1}
\docType{class}
\alias{OptCharacter-class}

\title{Class "OptCharacter"}
\description{Virtual class used internally for convenience.}
\section{Objects from the Class}{
  A virtual Class: No objects may be created from it.
}
\section{Methods}{
No methods defined with class "OptCharacter" in the signature.
}
\author{Andreas Alfons, \email{alfons@statistik.tuwien.ac.at}}
\examples{
showClass("OptCharacter")
}
\keyword{classes}
