\name{OptContControl-class}
\Rdversion{1.1}
\docType{class}
\alias{OptContControl-class}

\title{Class "OptContControl"}
\description{Virtual class used internally for convenience.}
\section{Objects from the Class}{
  A virtual Class: No objects may be created from it.
}
\section{Methods}{
No methods defined with class "OptContControl" in the signature.
}
\author{Andreas Alfons, \email{alfons@statistik.tuwien.ac.at}}
\seealso{
  \code{\linkS4class{SimControl}}
}
\examples{
showClass("OptContControl")
}
\keyword{classes}
