\name{OptDataControl-class}
\Rdversion{1.1}
\docType{class}
\alias{OptDataControl-class}
%% aliases to avoid confusion due to capitalization
\alias{optDataControl-class}
\alias{optdataControl-class}
\alias{optDatacontrol-class}
\alias{optdatacontrol-class}
\alias{OptdataControl-class}
\alias{Optdatacontrol-class}
\alias{OptDatacontrol-class}

\title{Class "OptDataControl"}
\description{Virtual class used internally for convenience.}
\section{Objects from the Class}{
  A virtual Class: No objects may be created from it.
}
\section{Methods}{
No methods defined with class \code{"OptDataControl"} in the signature.
}
\section{UML class diagram}{
  A slightly simplified UML class diagram of the framework can be found in 
  Figure 1 of the package vignette \emph{An Object-Oriented Framework for 
  Statistical Simulation: The \R Package \code{simFrame}}.  Use 
  \code{vignette("simFrame-intro")} to view this vignette.
}
\author{Andreas Alfons}
\seealso{
  \code{"\linkS4class{SimResults}"}
}
\examples{
showClass("OptDataControl")
}
\keyword{classes}
