\name{OptNAControl-class}
\Rdversion{1.1}
\docType{class}
\alias{OptNAControl-class}

\title{Class "OptNAControl"}
\description{Virtual class used internally for convenience.}
\section{Objects from the Class}{
  A virtual Class: No objects may be created from it.
}
\section{Methods}{
No methods defined with class "OptNAControl" in the signature.
}
\author{Andreas Alfons, \email{alfons@statistik.tuwien.ac.at}}
\seealso{
  \code{\linkS4class{SimControl}}
}
\examples{
showClass("OptNAControl")
}
\keyword{classes}
