\name{OptNumeric-class}
\Rdversion{1.1}
\docType{class}
\alias{OptNumeric-class}

\title{Class "OptNumeric"}
\description{Virtual class used internally for convenience.}
\section{Objects from the Class}{
  A virtual Class: No objects may be created from it.
}
\section{Methods}{
No methods defined with class "OptNumeric" in the signature.
}
\author{Andreas Alfons, \email{alfons@statistik.tuwien.ac.at}}
\examples{
showClass("OptNumeric")
}
\keyword{classes}
