\name{OptSampleControl-class}
\Rdversion{1.1}
\docType{class}
\alias{OptSampleControl-class}
%% aliases to avoid confusion due to capitalization
\alias{optSampleControl-class}
\alias{optsampleControl-class}
\alias{optSamplecontrol-class}
\alias{optsamplecontrol-class}
\alias{OptsampleControl-class}
\alias{Optsamplecontrol-class}
\alias{OptSamplecontrol-class}

\title{Class "OptSampleControl"}
\description{Virtual class used internally for convenience.}
\section{Objects from the Class}{
  A virtual Class: No objects may be created from it.
}
\section{Methods}{
No methods defined with class \code{"OptSampleControl"} in the signature.
}
\section{UML class diagram}{
  A slightly simplified UML class diagram of the framework can be found in 
  Figure 1 of the package vignette \emph{An Object-Oriented Framework for 
  Statistical Simulation: The \R Package \code{simFrame}}.  Use 
  \code{vignette("simFrame-intro")} to view this vignette.
}
\author{Andreas Alfons}
\seealso{
  \code{"\linkS4class{SimResults}"}
}
\examples{
showClass("OptSampleControl")
}
\keyword{classes}
