\name{SampleSetup}
\Rdversion{1.1}
\alias{SampleSetup}

\title{Create objects of class "SampleSetup"}
\description{Create objects of class \code{"SampleSetup"}.}
\usage{
SampleSetup(\dots)
}
\arguments{
    \item{\dots}{arguments passed to \code{new("SampleSetup", \dots)}.}
}
\value{
  An object of class \code{"SampleSetup"}.
}
\author{Andreas Alfons, \email{alfons@statistik.tuwien.ac.at}}
\seealso{
    \code{\linkS4class{SampleSetup}}
}
\keyword{classes}
