\name{SimResult}
\Rdversion{1.1}
\alias{SimResult}

\title{Create objects of class "SimResult"}
\description{Create objects of class \code{"SimResult"}.}
\usage{
SimResult(\dots)
}
\arguments{
    \item{\dots}{arguments passed to \code{new("SimResult", \dots)}.}
}
\value{
  An object of class \code{"SimResult"}.
}
\author{Andreas Alfons, \email{alfons@statistik.tuwien.ac.at}}
\seealso{
    \code{\linkS4class{SimResult}}
}
\keyword{classes}

