\name{SimResults}
\Rdversion{1.1}
\alias{SimResults}

\title{Create objects of class "SimResults"}
\description{Create objects of class \code{"SimResults"}.}
\usage{
SimResults(\dots)
}
\arguments{
    \item{\dots}{arguments passed to \code{new("SimResults", \dots)}.}
}
\value{
  An object of class \code{"SimResults"}.
}
\author{Andreas Alfons, \email{alfons@statistik.tuwien.ac.at}}
\seealso{
    \code{\linkS4class{SimResults}}
}
\keyword{classes}
