\name{Strata}
\Rdversion{1.1}
\alias{Strata}

\title{Create objects of class "Strata"}
\description{Create objects of class \code{"Strata"}.}
\usage{
Strata(\dots)
}
\arguments{
    \item{\dots}{arguments passed to \code{new("Strata", \dots)}.}
}
\value{
  An object of class \code{"Strata"}.
}
\author{Andreas Alfons, \email{alfons@statistik.tuwien.ac.at}}
\seealso{
    \code{\linkS4class{Strata}}
}
\keyword{classes}
