\name{stratify-methods}
\docType{methods}
\alias{stratify-methods}
\alias{stratify,data.frame,BasicVector-method}

\title{Methods for stratifying data}
\description{
Stratify data.
}
\section{Methods}{
\describe{

\item{x = "data.frame", design = "BasicVector"}{stratify data according to 
  the variables (columns) given by \code{design}.}
}}
\author{Andreas Alfons, \email{alfons@statistik.tuwien.ac.at}}
\seealso{
    \code{\link{stratify}}, \code{\linkS4class{Strata}}
}
\keyword{methods}
\keyword{manip}
