\name{OptCall-class}
\Rdversion{1.1}
\docType{class}
\alias{OptCall-class}

\title{Class \code{"OptCall"}}
\description{Virtual class used internally for convenience.}
\section{Objects from the Class}{
  A virtual Class: No objects may be created from it.
}
\section{Methods}{
No methods defined with class \code{"OptCall"} in the signature.
}
\author{Andreas Alfons}
\keyword{classes}
