\name{OptCall-class}
\Rdversion{1.1}
\docType{class}
\alias{OptCall-class}
%% aliases to avoid confusion due to capitalization
\alias{optCall-class}
\alias{optcall-class}
\alias{Optcall-class}

\title{Class "OptCall"}
\description{Virtual class used internally for convenience.}
\section{Objects from the Class}{
  A virtual Class: No objects may be created from it.
}
\section{Methods}{
No methods defined with class \code{"OptCall"} in the signature.
}
\author{Andreas Alfons}
\examples{
showClass("OptCall")
}
\keyword{classes}
