\name{simFrame-package}
\alias{simFrame-package}
\alias{simFrame}
%% aliases to avoid confusion due to capitalization
\alias{simframe-package}
\alias{SimFrame-package}
\alias{Simframe-package}
\alias{simframe}
\alias{SimFrame}
\alias{Simframe}

\docType{package}
\title{
\packageTitle{simFrame}
}
\description{
\packageDescription{simFrame}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{simFrame}
\packageIndices{simFrame}
}
\author{
Andreas Alfons [aut, cre]; C++ implementations of certain sampling algorithms are based on R code by Yves Tille and Alina Matei.

Maintainer: \packageMaintainer{simFrame}
}
\references{
Alfons, A., Templ, M. and Filzmoser, P. (2010) An Object-Oriented Framework for
Statistical Simulation: The \R Package \pkg{simFrame}. \emph{Journal of
Statistical Software}, \bold{37}(3), 1--36. \doi{10.18637/jss.v037.i03}.
}
\keyword{package}
